Complete refactor of views into several files to help organize code better
