+++ api/views/__init__.py
@@ -1,2 +1,3 @@
-from .views import *
 from .user_view import *
+from .forecast_view import *
+from .road_trip_view import *
